flask/requirements.txt
Flask==0.12.4
Werkzeug<0.13.0,>=0.12.0
google-cloud
google-cloud-storage
google-cloud-bigquery
requests-toolbelt
